reformat jsdoc comments, add test comments, test tweaks
