+++ package.json
@@ -22,7 +22,7 @@
     "@fortawesome/vue-fontawesome": "^3.0.0-5",
     "@sentry/tracing": "^7.9.0",
     "@sentry/vue": "^7.9.0",
-    "@vueuse/core": "^8.9.4",
+    "@vueuse/core": "^9.1.1",
     "apexcharts": "^3.35.4",
     "body-scroll-lock": "^4.0.0-beta.0",
     "core-js": "^3.23.5",
@@ -74,7 +74,7 @@
     "msw": "^0.44.2",
     "node-fetch": "^2.6.7",
     "prettier": "^2.7.1",
-    "prettier-plugin-jsdoc": "^0.3.38",
+    "prettier-plugin-jsdoc": "^0.4.2",
     "sass": "^1.53.0",
     "sass-loader": "^13.0.2",
     "ts-jest": "^27.0.0",
