+++ src/components/AppTable.vue
@@ -220,7 +220,8 @@ interface Props {
   search?: string;
   /**
    * whether to show certain controls (temp solution, needed b/c this is a
-   * controlled component and cannot paginate/search/etc on its own where needed yet)
+   * controlled component and cannot paginate/search/etc on its own where needed
+   * yet)
    */
   showControls?: boolean;
 }
