+++ src/global/meta.ts
@@ -3,7 +3,8 @@ import { ref, watch } from "vue";
 /**
  * index.html is statically generated with hard-coded meta data, from
  * environment variables specified in the .env file. convert these values into
- * reactive variables so we can change them and update the document tags during run time
+ * reactive variables so we can change them and update the document tags during
+ * run time
  */
 
 /** multi-part page title. array. gets joined with a | separator. */
