+++ tests/e2e/specs/node-search.cy.js
@@ -1,8 +1,10 @@
 it("Redirects to explore page from home page", () => {
+  /** go to homepage and focus search box */
   cy.visit("/");
   cy.get("input").trigger("focus");
   cy.location().should((loc) => expect(loc.pathname).to.eq("/explore"));
 
+  /** go to homepage and click one of tabs */
   cy.visit("/");
   cy.contains("Text Annotator").as("button").trigger("click");
   cy.location().should((loc) => expect(loc.pathname).to.eq("/explore"));
@@ -11,6 +13,7 @@ it("Redirects to explore page from home page", () => {
 it("Recent/frequent results show", () => {
   cy.visit("/explore");
 
+  /** dummy searches */
   const searches = [
     "abc def",
     "123",
@@ -21,17 +24,21 @@ it("Recent/frequent results show", () => {
     "abc def",
   ];
 
+  /** go through dummy searches */
   for (const search of searches) {
     cy.get("input").type(search);
+    /** dispatch textbox change event which triggers search and records it */
     cy.get("input").trigger("change");
     cy.wait(100);
     cy.get(".textbox button").trigger("click");
   }
 
+  /** go to node page, which should also get added to search history */
   cy.visit("/disease/MONDO:12345");
   cy.wait(100);
   cy.visit("/explore");
 
+  /** focus search box to show list of results */
   cy.get("input").trigger("focus", { force: true });
 
   /** recent */
@@ -45,6 +52,7 @@ it("Recent/frequent results show", () => {
 });
 
 it("Autocomplete results show", () => {
+  /** type something in search box for regular backend autocomplete results */
   cy.visit("/explore");
   cy.get("input").type("Marfan");
   cy.contains("Marfan and Marfan-related disorder");
@@ -55,6 +63,7 @@ it("Basic search results show", () => {
   cy.get("input").type("Marfan");
   cy.get("input").trigger("change");
 
+  /** search result with link shows */
   cy.contains("neonatal Marfan syndrome").as("result");
   cy.get("@result")
     .invoke("attr", "href")
@@ -66,6 +75,7 @@ it("Pagination works", () => {
   cy.get("input").type("Marfan");
   cy.get("input").trigger("change");
 
+  /** pagination text, and click through to next page */
   cy.contains("1 to 10 of 61 results");
   cy.contains("button", /^2$/).trigger("click");
   cy.contains("11 to 20 of 61 results");
@@ -76,9 +86,10 @@ it("Filters show", () => {
   cy.get("input").type("Marfan");
   cy.get("input").trigger("change");
 
+  /** filters show */
+  /** actual filtering done by backend, so not much to test here */
   cy.contains("Category").trigger("click");
   cy.contains(/Disease*.25/).should("exist");
-
   cy.contains("Taxon").trigger("click");
   cy.contains(/Gallus Gallus*.1/).should("exist");
 });
